requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.5.3
aiogram==3.4.1
greenlet
python-multipart